requirements.txt
websockets==12.0
py2app==0.27.0
pyobjc-core==9.2
pyobjc-framework-Cocoa==9.2
setuptools>=69.0.0
importlib-metadata>=7.0.0
packaging>=23.2 